pyproject.toml
[project]
name = "bot-gateway"
version = "0.1.0"
description = "Multi-platform message gateway for LLM backends."
readme = "CLAUDE.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "httpx",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "python-pptx>=1.0.2",
    "redis",
    "uvicorn",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
